arm_mptimer: Convert to use ptimer
Current ARM MPTimer implementation uses QEMUTimer for the actual timer,
this implementation isn't complete and mostly tries to duplicate of what
generic ptimer is already doing fine.

Conversion to ptimer brings the following benefits and fixes:
	- Simple timer pausing implementation
	- Fixes counter value preservation after stopping the timer
	- Code simplification and reduction

Bump timerblock VMSD to version 3, since VMState is changed and is not
compatible with the previuos implementation.

Signed-off-by: Dmitry Osipenko <digetx@gmail.com>